IGNORING HALL FAULTS
Hall faults are detected by the fpga code when we get a reading of 000 or 111
from a hall sensor.  When this happens, a bit is set in motor_faults for the
corresponding motor.  This seems to happen often to our robots due to crappy
motor disassembly/reassembly, especially when the motor back plates are mounted
at a slight tilt.

We still register motor/hall faults as before, but ignore them in the control
systems on the bot, so they continue to function normally

This is a change in robot firmware and requires every control board to
be re-flashed.