fixed issue with googletest
We had CMake setup to clone googletest from the submodule in the /external dir,
which defaulted to taking whatever was on the master branch.  B/c of the way
submodules work, the "master" label isnt set to anything in particular, it just
stores the commit hash, which meant that different machines would build a
different version of googletest.  this was fixed by using the URL
specifier for ExternalProject instead of GIT_REPO