fixed calibration issue, transmitter now runs forever
previous to these changes, the sender-test program would only be able
to send 4 or 5 packets before the cc debugger would stop getting anything
from it.  the sender could be rebooted and be able to send another 5, but
then would fail to send meaningful data.

this commit changes the getData() method to only calibrate the radio
(via SAFC strobe) if the packet was received successfully. this keeps us
from calling SAFC when the getData() method is triggered via a transmit