Stop emitting modelReset
This is a private signal, so we shouldn't ever be emitting it.

I'm not sure what this was originally used for, so let me know if you
see any regressions.

From my tests, this dosen't change anything, except that the play tree
no longer collapses when you load a playbook