[MP Rewrite Part 2] Introduce the Trajectory abstraction (#1494)
A Trajectory is a piecewise-cubic function interpolating between a
vector of time-indexed RobotInstant waypoints.

This is not expected to pass CI.

* Add motion control changes for following Planning::Trajectory

* Address review comments

* Change fixAngleRadians to be more efficient

* Address review comments

* Tidy up != operators (use == in defn.)

* [MP Rewrite Part 3] Add PlanRequest and the planner node. (#1495)

* Add PlanRequest and the planner node.

This also overhauls the MotionCommand structure.

* Remove dead code from Processor, replace with TODO

* Make path target goal a linear motion instant

* Address review comments

[MP Rewrite Part 4] Motion Planning Primitives (#1496)

Primitives for motion planning including angle planning, RRT, velocity
profiling, Bezier fitting, and replanning. This is anticipated to be the
largest portion of the motion planning rewrite.

This is not expected to pass CI.

* Remove disabled for-loop in Processor

* Fix TODOs and docs in planning primitives

* Address review comments

* Make off-path threshold configurable

[MP Rewrite Part 5] Port previous planners to new framework (#1497)

Most of these are not functional changes from the old planners, but
rather simple ports to the new primitives. The biggest changes are that
we now use `Replanner` static functions and manually manage our own
"previous trajectory" objects instead of having "behavior-level"
planners (i.e. settle) own instances of an RRT planner and interacting
with it through plan requests.

This is the final PR in the rewrite, and it should pass CI. At the point
where it passes CI and all PRs are ready for merge, each PR should be
merged into its parent in sequence.

* Address review comments

* Move applyLowPassFilter into common, add TODOs

* Enable escape-obstacles planner

[MP Rewrite Part 6] Cleanup and fixes (#1512)

* Cleanup and fixes to make code run properly

* Fix test compilation and disable a trapezoidal motion profile edge case.

Some planner tests are still failing. This will be addressed in a later
commit.

* Prettify

* Make e2e planner tests pass.

Because they generate random scenarios, and the RRT planner is not
guaranteed to converge in a finite number of iterations (and the Bezier
approximation may be incorrect), these tests are currently flaky. That
should be fixed before this is merged to staging.

* Tidy fixes

* Linting

* Formatting

* Remove deleted file from CMakeLists, remove reuse condition from settle

* Don't replan in Collect::control

* Optimizations to geometry methods

* Reset and regenerate sim config

This deletes unnecessary variables and moves planning-related variables
into PathPlanner

* Formatting

* Add TODO in EscapeObstacles

* Tidy fixes

* Revert inconsistent clang-tidy recommended fix

* Temporarily disable flaky planner test

* Add NOLINTNEXTLINE

* Remove old sim config and add TODOs for trajectory utils