Motion Planning Rewrite (#1492)
* Deleted to-be-removed files in motion planning.

[MP Rewrite Part 1] Break out and clean up BallState (#1493)

This substantially cleans up the ball state code, and moves it from
`SystemState` (which we are moving away from in general) to `WorldState`
(which is restricted to physics-only information.

[MP Rewrite Part 2] Introduce the Trajectory abstraction (#1494)

A Trajectory is a piecewise-cubic function interpolating between a
vector of time-indexed RobotInstant waypoints.

[MP Rewrite Part 3] Add PlanRequest and the planner node (#1495)

This also overhauls the MotionCommand structure.

[MP Rewrite Part 4] Motion Planning Primitives (#1496)

Primitives for motion planning including angle planning, RRT, velocity
profiling, Bezier fitting, and replanning. This is anticipated to be the
largest portion of the motion planning rewrite.

[MP Rewrite Part 5] Port previous planners to new framework (#1497)

Most of these are not functional changes from the old planners, but
rather simple ports to the new primitives. The biggest changes are that
we now use `Replanner` static functions and manually manage our own
"previous trajectory" objects instead of having "behavior-level"
planners (i.e. settle) own instances of an RRT planner and interacting
with it through plan requests.

[MP Rewrite Part 6] Cleanup and fixes (#1512)

Because they generate random scenarios, and the RRT planner is not
guaranteed to converge in a finite number of iterations (and the Bezier
approximation may be incorrect), these tests are currently flaky. That
should be fixed before this is merged to staging.

* Reset and regenerate sim config. This deletes unnecessary variables and moves planning-related variables
into PathPlanner

[MP Rewrite Part 7] Add global obstacles; fix dynamic obstacle properties (#1526)

[MP Rewrite Part 8] Optimize geometry checking (#1534)

[MP Rewrite Part 9] Improve test coverage for motion planning primitives (#1536)

[MP Rewrite Part 10] Overall improvements to planners (#1537)

* Add angle overrides and fix pivot planner.

* Make planner tests deterministic

* Make settle planner pick best buffer

This fixes the edge case where the planner can't find an path that finds
the ball quickly enough because the only way to catch it doesn't have
enough buffer time (i.e. immediately before it is caught).

[MP Rewrite Part 11] Miscellaneous formatting and cleanup (#1540)

[MP Rewrite Part 12] Port the line kick planner (#1541)

[MP Rewrite Part 13] Stop play fixes and holding trajectory (#1542)